LayoutTests/fast/forms/validation-message-on-checkbox-expected.txt

Check if the position of a validation message for a checkbox is good.


The top of the bubble should be on the bottom of the host node:
PASS bubbleRect.top is hostRect.bottom

The left position of the bubble should be 0 because of no enough space:
PASS hostRect.left + hostRect.width / 2 < 32 is true
PASS bubbleRect.left is 0

PASS successfullyParsed is true

TEST COMPLETE
